+++ services/order-processing/src/main/java/com/eshop/ordering/api/application/integrationevents/events/models/ConfirmedOrderStockItem.java
@@ -1,4 +1,4 @@
 package com.eshop.ordering.api.application.integrationevents.events.models;
 
-public record ConfirmedOrderStockItem(Integer productId, boolean hasStock) {
+public record ConfirmedOrderStockItem(Long productId, boolean hasStock) {
 }
